# services/auth-service/openapi.yaml
openapi: 3.0.3
info:
  title: GameTriq Authentication Service
  description: |
    Basketball League Management Platform Authentication Service with COPPA compliance.
    
    This service provides secure authentication, authorization, and user management
    for the GameTriq basketball league platform, with special considerations for
    youth sports participants under 13 years old (COPPA compliance).
    
    ## Key Features
    - JWT-based authentication with refresh tokens
    - Multi-factor authentication (MFA) for eligible users
    - COPPA-compliant registration and parental consent workflows
    - Age-appropriate password policies
    - Multi-tenant architecture support
    - Session management with automatic expiration
    - Comprehensive audit logging
    
    ## Security Considerations
    - All passwords are hashed using bcrypt with salt rounds
    - JWT tokens have short expiration times (15 minutes for access tokens)
    - Rate limiting is applied to prevent brute force attacks
    - COPPA compliance includes parental consent verification
    - Enhanced data protection for users under 13
  version: 1.0.0
  contact:
    name: GameTriq Development Team
    email: dev@gametriq.com
  license:
    name: Proprietary
    url: https://gametriq.com/license
servers:
  - url: https://auth.gametriq.com/api/v1
    description: Production server
  - url: https://auth-staging.gametriq.com/api/v1
    description: Staging server
  - url: http://localhost:3001/api/v1
...
paths:
...
  /refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate a new access token using a valid refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/change-password:
    post:
      tags:
        - User Management
      summary: Change user password
      description: Change the current user's password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Current password incorrect or authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/mfa/enable:
    post:
      tags:
        - MFA
      summary: Enable multi-factor authentication
      description: |
        Enable MFA for the current user. Not available for users under 13.
        Returns QR code and backup codes for setup.
      operationId: enableMFA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableMFARequest'
      responses:
        '200':
          description: MFA enabled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFASetupResponse'
        '401':
          description: Authentication required or password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: MFA not available for users under 13
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/mfa/verify:
    post:
      tags:
        - MFA
      summary: Verify MFA token
      description: Verify a TOTP token or backup code
      operationId: verifyMFA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyMFARequest'
      responses:
        '200':
          description: MFA token verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MFAVerificationResponse'
  /parental-consent:
    post:
      tags:
        - COPPA Compliance
      summary: Request parental consent
      description: |
        Initiate parental consent process for a user under 13.
        Sends verification email or initiates other verification methods.
      operationId: requestParentalConsent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentalConsentRequest'
      responses:
        '200':
          description: Parental consent request initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParentalConsentResponse'
        '400':
          description: Invalid request or child not eligible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Child user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /parental-consent/{consentId}/verify:
    post:
      tags:
        - COPPA Compliance
      summary: Verify parental consent
      description: Verify parental consent using provided verification data
      operationId: verifyParentalConsent
      parameters:
        - name: consentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Consent request ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentVerificationRequest'
      responses:
        '200':
          description: Consent verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConsentVerificationResponse'
        '400':
          description: Verification failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login endpoint
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          description: Valid refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: Current password
          example: "OldPass123!"
        newPassword:
          type: string
          minLength: 8
          description: New password (minimum 8 characters)
          example: "NewSecurePass123!"
    EnableMFARequest:
      type: object
      required:
        - password
      properties:
        password:
          type: string
          description: User's current password for verification
          example: "SecurePass123!"
    VerifyMFARequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: 6-digit TOTP token or backup code
          example: "123456"
...
    ConsentVerificationRequest:
      type: object
      properties:
        token:
          type: string
          description: Email verification token
        signature:
          type: string
          description: Digital signature data
        signatureValid:
          type: boolean
          description: Digital signature validation result
        authorizationCode:
          type: string
          description: Credit card authorization code
        verified:
          type: boolean
          description: Credit card verification result
        idVerified:
          type: boolean
          description: Government ID verification result
        confidence:
          type: number
          description: ID verification confidence score
        staffVerified:
          type: boolean
          description: Staff verification result for video calls
        staffId:
          type: string
          description: ID of staff member who verified

    # Response Schemas
...
    TokenResponse:
      type: object
      required:
        - success
        - message
        - tokens
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Token refreshed successfully"
        tokens:
          $ref: '#/components/schemas/TokenPair'
...
    MFASetupResponse:
      type: object
      required:
        - success
        - message
        - mfa
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "MFA enabled successfully"
        mfa:
          type: object
          required:
            - secret
            - qrCodeUrl
            - backupCodes
          properties:
            secret:
              type: string
              description: Base32 encoded secret for TOTP
              example: "JBSWY3DPEHPK3PXP"
            qrCodeUrl:
              type: string
              format: uri
              description: QR code image as data URL
              example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAQ..."
            backupCodes:
              type: array
              items:
                type: string
              description: 8 backup codes for account recovery
              example: ["12345678", "87654321", "11223344", "44332211", "55667788", "88776655", "99001122", "22110099"]
    MFAVerificationResponse:
      type: object
      required:
        - success
        - valid
        - message
      properties:
        success:
          type: boolean
          example: true
        valid:
          type: boolean
          description: Whether the MFA token is valid
          example: true
        message:
          type: string
          example: "MFA token is valid"
    ParentalConsentResponse:
      type: object
      required:
        - success
        - message
        - consent_id
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Parental consent request initiated"
        consent_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        verification_url:
          type: string
          format: uri
          description: URL for parent to complete verification
          example: "https://gametriq.com/coppa/verify?consent=123e4567-e89b-12d3-a456-426614174000&token=abc123"
...
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type
          example: "Validation Error"
        message:
          type: string
          description: Error message
          example: "The provided email address is invalid"
        timestamp:
          type: string
          format: date-time
          example: "2023-12-07T10:30:00Z"

    # Common Data Types
...
security:
  - BearerAuth: []